actualización ejercicios 7 y 8

+++ src/ejercicio5.py
@@ -1,4 +1,4 @@
-imp = int(input("Introduce el importe sin IVA: "))
-iva = int(input("Introduce el tipo de IVA:  "))
+imp = float(input("Introduce el importe sin IVA: "))
+iva = float(input("Introduce el tipo de IVA:  "))
 
 print("El precio total + IVA es: ", imp + (imp * iva / 100))
